requirements.txt
streamlit
pdfplumber
python-docx
spacy>=3.7.0,<3.8.0
numpy<2.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
sentence-transformers